requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==20.8b1
click==7.1.2
dj-database-url==0.5.0
Django==3.2
django-common-helpers==0.9.2
django-crontab==0.7.0
djangorestframework==3.12.4
flake8==3.9.0
gunicorn==20.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.0
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
whitenoise==5.2.0
